Refactor Networking into Network package
