+++ network/networks_test.go
@@ -13,7 +13,7 @@ func TestAccNetworks_List(t *testing.T) {
 		Steps: []testutils.Step{
 			&testutils.StepAPICall{
 				StateBagKey: "networks",
-				CallFunc: func(client *Networking) (interface{}, error) {
+				CallFunc: func(client *NetworkService) (interface{}, error) {
 					return client.Networks().ListNetworks(
 						context.Background(),
 						&ListNetworksInput{})
